shift-fix display user filter. recipe-add comment

+++ r_edit_update.php
@@ -5,6 +5,7 @@
 	version: 0,
 	recipe: 0,
 	cat: "",
+  comment: "",
 	arrayItemM: [], //recipe item
 	arrayItemU: [], //recipe unit
 	arrayItemQ: [], //recipe qty
@@ -31,8 +32,7 @@
   $c_version = $obj->version;
   $c_recipe = $obj->recipe;
   $c_cat = $obj->cat;
-
-  myLOG($obj);
+  $c_comment = $obj->comment;
 
   try{
     $conn->autocommit(false);
@@ -46,7 +46,7 @@
       $sql = "UPDATE `t_config` SET `c_value`=".($c_recipe+1)." WHERE `c_setup`='recipe_num'";
       $result = $conn->query($sql);
       if (!$result) throw new Exception('Error updating new recipe number!');
-      $sql = "INSERT INTO `t_recipe` (`c_recipe`,`c_product`,`c_cat`,`c_version`) VALUES (".$c_recipe.",'".$c_product."','".$c_cat."',$c_version)";
+      $sql = "INSERT INTO `t_recipe` (`c_recipe`,`c_product`,`c_cat`,`c_version`,`c_comment`) VALUES (".$c_recipe.",'".$c_product."','".$c_cat."',".$c_version.",'".$c_comment."')";
       $result = $conn->query($sql);
       if (!$result) throw new Exception('Error creating new product!');
       $totalRow = count($obj->arrayItemM);
@@ -60,6 +60,12 @@
         if (!$result) throw new Exception('Error creating new recipe item!');
       }
     }else{//update existing recipe
+      //update comment field
+      $sql = "UPDATE `t_recipe` SET `c_comment`='".$c_comment."' WHERE `c_recipe`=".$c_recipe;
+      $result = $conn->query($sql);
+      if (!$result){
+        throw new Exception('Error updating recipe comment!');
+      }
       $sql = "DELETE FROM `t_recipelib` WHERE `c_recipe`=".$c_recipe;
       $result = $conn->query($sql);
       if ($result){
@@ -70,7 +76,6 @@
           $c_quantity = $obj->arrayItemQ[$i];
           $c_base = $obj->arrayItemB[$i];
           $sql = "INSERT INTO `t_recipelib` (`c_recipe`,`c_material`,`c_quantity`,`c_unit`,`c_base`) VALUES (".$c_recipe.",'".$c_material."',".$c_quantity.",'".$c_unit."',".$c_base.")";
-          myLOG($sql);
           $result = $conn->query($sql);
           if (!$result) throw new Exception('Error creating new recipe when updating recipe with item:'.$c_material."!");
         }
@@ -84,7 +89,6 @@
       throw new Exception('Error commit database changes!');
     }
   } catch (Exception $e){
-    myLOG("rollback");
     $conn->rollback();
     echo json_encode($e->getMessage());
   }
